fud::dht

Trait DhtHandler

pub trait DhtHandler {
    // Required methods
    fn dht(&self) -> Arc<Dht>;
    fn ping<'life0, 'async_trait>(
        &'life0 self,
        channel: ChannelPtr,
    ) -> Pin<Box<dyn Future<Output = Result<DhtNode>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn on_new_node<'life0, 'life1, 'async_trait>(
        &'life0 self,
        node: &'life1 DhtNode,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn fetch_nodes<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        node: &'life1 DhtNode,
        key: &'life2 Hash,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<DhtNode>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;

    // Provided methods
    fn announce<'life0, 'life1, 'life2, 'async_trait, M>(
        &'life0 self,
        key: &'life1 Hash,
        message: &'life2 M,
        router: Arc<RwLock<HashMap<Hash, HashSet<DhtRouterItem>>>>,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where M: 'async_trait + Message,
             Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn channel_task<'life0, 'async_trait, M>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where M: 'async_trait + Message,
             Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn disconnect_task<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn add_node<'life0, 'async_trait>(
        &'life0 self,
        node: DhtNode,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn update_node<'life0, 'life1, 'async_trait>(
        &'life0 self,
        node: &'life1 DhtNode,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn lookup_nodes<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 Hash,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<DhtNode>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn get_channel<'life0, 'life1, 'async_trait>(
        &'life0 self,
        node: &'life1 DhtNode,
    ) -> Pin<Box<dyn Future<Output = Result<ChannelPtr>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn add_to_router<'life0, 'life1, 'async_trait>(
        &'life0 self,
        router: Arc<RwLock<HashMap<Hash, HashSet<DhtRouterItem>>>>,
        key: &'life1 Hash,
        router_items: Vec<DhtRouterItem>,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Required Methods

fn dht(&self) -> Arc<Dht>

fn ping<'life0, 'async_trait>( &'life0 self, channel: ChannelPtr, ) -> Pin<Box<dyn Future<Output = Result<DhtNode>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn on_new_node<'life0, 'life1, 'async_trait>( &'life0 self, node: &'life1 DhtNode, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn fetch_nodes<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, node: &'life1 DhtNode, key: &'life2 Hash, ) -> Pin<Box<dyn Future<Output = Result<Vec<DhtNode>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Provided Methods

fn announce<'life0, 'life1, 'life2, 'async_trait, M>( &'life0 self, key: &'life1 Hash, message: &'life2 M, router: Arc<RwLock<HashMap<Hash, HashSet<DhtRouterItem>>>>, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where M: 'async_trait + Message, Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn channel_task<'life0, 'async_trait, M>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where M: 'async_trait + Message, Self: Sync + 'async_trait, 'life0: 'async_trait,

fn disconnect_task<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

fn add_node<'life0, 'async_trait>( &'life0 self, node: DhtNode, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

fn update_node<'life0, 'life1, 'async_trait>( &'life0 self, node: &'life1 DhtNode, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Move a node to the tail in its bucket, to show that it is the most recently seen in the bucket. If the node is not in a bucket it will be added using add_node

fn lookup_nodes<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 Hash, ) -> Pin<Box<dyn Future<Output = Result<Vec<DhtNode>>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_channel<'life0, 'life1, 'async_trait>( &'life0 self, node: &'life1 DhtNode, ) -> Pin<Box<dyn Future<Output = Result<ChannelPtr>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn add_to_router<'life0, 'life1, 'async_trait>( &'life0 self, router: Arc<RwLock<HashMap<Hash, HashSet<DhtRouterItem>>>>, key: &'life1 Hash, router_items: Vec<DhtRouterItem>, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.