fud

Module download

Download methods

Macros

seeders_loop 🔒

Receive seeders from a DHT events subscription, and execute an async expression for each deduplicated seeder once (seeder order is random). It will keep going until the expression returns Ok(()), or there are no more seeders. It has an optional favored_seeder argument that will be tried first if specified.

Structs

ChunkFetchContext 🔒

Enums

ChunkFetchControl 🔒

MetadataFetchReply 🔒

Functions

fetch_chunk 🔒

Fetch a single chunk and return what should be done next

fetch_chunks

Fetch chunks for chunked (file or directory) from seeders in seeders_sub.

fetch_metadata

Fetch a single resource metadata from seeders received from seeders_sub. If the resource is a file smaller than a single chunk then seeder can send the chunk directly, and we will create the file from it on path path.

handle_chunk_reply 🔒

Processes an incoming chunk

Type Aliases

FudDhtEvent 🔒