fud

Module equix

Equi-X Proof-of-Work using Equi-X

https://spec.torproject.org/hspow-spec/v1-equix.html https://github.com/tevador/equix/blob/master/devlog.md https://github.com/tevador/equix

Structs

Challenge

A challenge string

EquiXBuilder

Builder for creating [EquiX] instances with custom settings

EquiXPow

Solution

Potential solution to an EquiX puzzle

SolverMemory

Temporary memory used by the Equi-X solver

Enums

HashError

Errors that could occur while building a hash function

RuntimeOption

Option for selecting a HashX runtime

Constants

NONCE_LEN

Length of the nonce value generated by clients and included in the solution

P_STRING 🔒

Algorithm personalization string

P_STRING_LEN 🔒

Length of the personalization string, in bytes