fud

Struct Fud

pub struct Fud {

    pub(crate) state: Arc<RwLock<Option<FudState>>>,
    pub(crate) geode: Geode,
    pub(crate) downloads_path: PathBuf,
    pub(crate) chunk_timeout: u64,
    pub pow: Arc<RwLock<FudPow>>,
    pub(crate) dht: Arc<Dht<Fud>>,
    pub(crate) resources: Arc<RwLock<HashMap<Hash, Resource>>>,
    pub(crate) path_tree: Tree,
    pub(crate) file_selection_tree: Tree,
    pub(crate) scrap_tree: Tree,
    pub(crate) get_tx: Sender<(Hash, PathBuf, FileSelection)>,
    pub(crate) get_rx: Receiver<(Hash, PathBuf, FileSelection)>,
    pub(crate) put_tx: Sender<PathBuf>,
    pub(crate) put_rx: Receiver<PathBuf>,
    pub(crate) lookup_tx: Sender<Hash>,
    pub(crate) lookup_rx: Receiver<Hash>,
    pub(crate) verify_node_tx: Sender<FudNode>,
    pub(crate) verify_node_rx: Receiver<FudNode>,
    pub(crate) fetch_tasks: Arc<RwLock<HashMap<Hash, Arc<StoppableTask>>>>,
    pub(crate) put_tasks: Arc<RwLock<HashMap<PathBuf, Arc<StoppableTask>>>>,
    pub(crate) lookup_tasks: Arc<RwLock<HashMap<Hash, Arc<StoppableTask>>>>,
    pub(crate) tasks: Arc<RwLock<HashMap<String, Arc<StoppableTask>>>>,
    pub(crate) event_publisher: PublisherPtr<FudEvent>,
    pub(crate) p2p: P2pPtr,
    pub executor: ExecutorPtr,

}

Fields
state: Arc<RwLock<Option<FudState>>>geode: Geode

The Geode instance

downloads_path: PathBuf

Default download directory

chunk_timeout: u64

Chunk transfer timeout in seconds

pow: Arc<RwLock<FudPow>>

The FudPow instance

dht: Arc<Dht<Fud>>

The DHT instance

resources: Arc<RwLock<HashMap<Hash, Resource>>>

Resources (current status of all downloads/seeds)

path_tree: Tree

Sled tree containing “resource hash -> path on the filesystem”

file_selection_tree: Tree

Sled tree containing “resource hash -> file selection”. If the file selection is all files of the resource (or if the resource is not a directory), the resource does not store its file selection in the tree.

scrap_tree: Tree

Sled tree containing scraps which are chunks containing data the user did not want to save to files. They also contain data the user wanted otherwise we would not have downloaded the chunk at all. We save scraps to be able to verify integrity even if part of the chunk is not saved to the filesystem in the downloaded files. “chunk/scrap hash -> chunk content”

get_tx: Sender<(Hash, PathBuf, FileSelection)>

Get requests sender

get_rx: Receiver<(Hash, PathBuf, FileSelection)>

Get requests receiver

put_tx: Sender<PathBuf>

Put requests sender

put_rx: Receiver<PathBuf>

Put requests receiver

lookup_tx: Sender<Hash>

Lookup requests sender

lookup_rx: Receiver<Hash>

Lookup requests receiver

verify_node_tx: Sender<FudNode>

Verify node requests sender

verify_node_rx: Receiver<FudNode>

Verify node requests receiver

fetch_tasks: Arc<RwLock<HashMap<Hash, Arc<StoppableTask>>>>

Currently active downloading tasks (running the fud.fetch_resource() method)

put_tasks: Arc<RwLock<HashMap<PathBuf, Arc<StoppableTask>>>>

Currently active put tasks (running the fud.insert_resource() method)

lookup_tasks: Arc<RwLock<HashMap<Hash, Arc<StoppableTask>>>>

Currently active lookup tasks (running the fud.lookup_value() method)

tasks: Arc<RwLock<HashMap<String, Arc<StoppableTask>>>>

Currently active tasks (defined in tasks, started with the start_task macro)

event_publisher: PublisherPtr<FudEvent>

Used to send events to fud clients

p2p: P2pPtr

Pointer to the P2P network instance

executor: ExecutorPtr

Global multithreaded executor reference

Implementations

impl Fud

pub async fn new( settings: Args, p2p: P2pPtr, sled_db: &Db, event_publisher: PublisherPtr<FudEvent>, executor: ExecutorPtr, ) -> Result<Arc<Self>>

pub async fn start(self: &Arc<Self>) -> Result<()>

Run the PoW and generate a VerifiableNodeData, then start tasks

pub(crate) async fn start_tasks(self: &Arc<Self>)

pub(crate) async fn init(&self) -> Result<()>

Verify our resources, add ourselves to the seeders (dht.hash_table) for the resources we already have, announce our resources.

pub async fn resources(&self) -> HashMap<Hash, Resource>

Get a copy of the current resources

pub fn hash_to_path(&self, hash: &Hash) -> Result<Option<PathBuf>>

Get resource path from hash using the sled db

pub fn path_to_hash(&self, path: &Path) -> Result<Option<Hash>>

Get resource hash from path using the sled db

pub async fn new_seeder(&self, key: &Hash) -> Result<FudSeeder>

Create a new dht::FudSeeder for own node

pub async fn verify_resources( &self, hashes: Option<Vec<Hash>>, ) -> Result<Vec<Resource>>

Verify if resources are complete and uncorrupted. If a resource is incomplete or corrupted, its status is changed to Incomplete. If a resource is complete, its status is changed to Seeding. Takes an optional list of resource hashes. If no hash is given (None), it verifies all resources. Returns the list of verified and uncorrupted/complete seeding resources.

pub async fn get( &self, hash: &Hash, path: &Path, files: FileSelection, ) -> Result<()>

Start downloading a file or directory from the network to path. This creates a new task in fetch_tasks calling fetch_resource(). files is the list of files (relative paths) you want to download (if the resource is a directory), None means you want all files.

pub async fn get_metadata( &self, hash: &Hash, path: &Path, ) -> Result<(ChunkedStorage, Option<FudSeeder>)>

Try to get the chunked file or directory from geode, if we don’t have it then it is fetched from the network using fetch_metadata(). If we need to fetch from the network, the seeders we find are sent to seeders_pub. The seeder in the returned result is only defined if we fetched from the network.

pub async fn fetch_resource( &self, hash: &Hash, path: &Path, files: &FileSelection, ) -> Result<()>

Download a file or directory from the network to path. Called when get() creates a new fetch task.

pub(crate) async fn write_scraps( &self, chunked: &mut ChunkedStorage, chunk_hashes: &HashSet<Hash>, ) -> Result<()>

pub async fn verify_chunks( &self, resource: &Resource, chunked: &mut ChunkedStorage, ) -> Result<(u64, u64)>

Iterate over chunks and find which chunks are available locally, either in the filesystem (using geode::verify_chunks()) or in scraps. chunk_hashes is the list of chunk hashes we want to take into account, None means to take all chunks into account. Return the scraps in a HashMap, and the size in bytes of locally available data (downloaded and downloaded+targeted).

pub async fn put(&self, path: &Path) -> Result<()>

Add a resource from the file system.

pub async fn insert_resource(&self, path: &PathBuf) -> Result<()>

Insert a file or directory from the file system. Called when put() creates a new put task.

pub async fn remove(&self, hash: &Hash)

Removes:

a resource

its metadata in geode

its path in the sled path tree

its file selection in the sled file selection tree

and any related scrap in the sled scrap tree,

then sends a ResourceRemoved fud event.

pub async fn prune_seeders(&self, expiry_secs: u32)

Remove seeders that are older than expiry_secs

pub async fn stop(&self)

Stop all tasks.