fud

Module util

Utils

Enums

FileSelection

An enum to represent a set of files, where you can use All if you want all files without having to specify all of them. We could use an Option<HashSet<PathBuf>>, but this is more explicit.

Functions

create_all_files

get_all_files

hash_to_string

receive_resource_msg

Wait for a crate::proto::ResourceMessage on msg_subscriber with a timeout. If we receive a message with the wrong resource hash, it’s skipped.