fud

Module tasks

Enums

FetchReply

Functions

announce_seed_task

Background task that announces our files and chunks once every hour. Also removes seeders that did not announce for too long.

fetch_file_task

Background task that receives file fetch requests and tries to fetch objects from the network using the routing table. TODO: This can be optimised a lot for connection reuse, etc.

get_task

Triggered when calling the get RPC method