fud

Module tasks

Background tasks

Macros

start_task 🔒

Functions

announce_seed_task

Background task that announces our files once every hour. Also removes seeders that did not announce for too long.

get_task

Triggered when calling the fud.get() method. It creates a new StoppableTask (running fud.fetch_resource()) and inserts it into the fud.fetch_tasks hashmap. When the task is stopped it’s removed from the hashmap.

handle_dht_events

Handle DHT events in fud.

lookup_task

Triggered when you need to lookup seeders for a resource.

node_id_task

Background task that:

put_task

Triggered when calling the fud.put() method.

verify_node_task

After pinging an inbound connection, this task is triggered to make sure that you are able to reach at least one of the node’s external address. Fud::ping() will take care of adding the node to our buckets.